+++ App/App_macOS.swift
@@ -67,6 +67,8 @@ struct Kiwix: App {
                 .environmentObject(libraryRefreshViewModel)
                 .task { colorSchemeStore.update() }
                 .modifier(OpeningSettingsModifier(updateTabSelection: selectHotspotTab))
+                .modifier(ValidationModifier())
+                .disabled(ValidationShared.state.isValidating)
         }.commands {
             SidebarCommands()
             CommandGroup(replacing: .importExport) {
@@ -129,6 +131,7 @@ struct Kiwix: App {
                     .tag(SettingsTab.about.rawValue)
                 if FeatureFlags.hasLibrary {
                     DiagnosticsView()
+                        .environment(\.managedObjectContext, Database.shared.viewContext)
                         .tag(SettingsTab.diagnostics.rawValue)
                 }
             }